refactor: move test helpers to their own files
Closes #987
